[#13] Added types directory
